modularize client tests
