+++ generators/client/generator.spec.mjs
@@ -18,19 +18,26 @@
  */
 import { expect } from 'expect';
 import lodash from 'lodash';
-import { basename, dirname } from 'path';
+import { basename, dirname, join } from 'path';
 import { fileURLToPath } from 'url';
+import assert from 'yeoman-assert';
 
 import testSupport from '../../test/support/index.cjs';
 import Generator from './index.js';
+import { skipPrettierHelpers as helpers } from '../../test/utils/utils.mjs';
+import constants from '../generator-defaults.js';
+import TestFrameworkTypes from '../../jdl/jhipster/test-framework-types.js';
 
 const { snakeCase } = lodash;
 const { testBlueprintSupport } = testSupport;
+const { appDefaultConfig } = constants;
+const { CYPRESS } = TestFrameworkTypes;
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
 const generator = basename(__dirname);
+const generatorFile = join(__dirname, 'index.js');
 
 describe(`JHipster ${generator} generator`, () => {
   it('generator-list constant matches folder name', async () => {
@@ -44,4 +51,126 @@ describe(`JHipster ${generator} generator`, () => {
     expect(instance.features.bar).toBe(true);
   });
   describe('blueprint support', () => testBlueprintSupport(generator));
+
+  describe('composing', () => {
+    const mockedComposedGenerators = ['jhipster:common', 'jhipster:client', 'jhipster:languages', 'jhipster:cypress'];
+
+    describe('with translation disabled', () => {
+      let runResult;
+      const options = { enableTranslation: false };
+      before(() => {
+        return helpers
+          .create(generatorFile)
+          .withOptions({
+            fromCli: true,
+            skipInstall: true,
+            defaultLocalConfig: { ...appDefaultConfig, ...options },
+          })
+          .withMockedGenerators(mockedComposedGenerators)
+          .run()
+          .then(result => {
+            runResult = result;
+          });
+      });
+
+      after(() => runResult.cleanup());
+
+      it('should compose with jhipster:common', () => {
+        assert(runResult.mockedGenerators['jhipster:common'].calledOnce);
+      });
+      it('should not compose with jhipster:languages', () => {
+        assert.equal(runResult.mockedGenerators['jhipster:languages'].callCount, 0);
+      });
+    });
+
+    describe('with translation enabled', () => {
+      let runResult;
+      const options = { enableTranslation: true };
+      before(() => {
+        return helpers
+          .create(generatorFile)
+          .withOptions({
+            fromCli: true,
+            skipInstall: true,
+            defaultLocalConfig: { ...appDefaultConfig, ...options },
+          })
+          .withMockedGenerators(mockedComposedGenerators)
+          .run()
+          .then(result => {
+            runResult = result;
+          });
+      });
+
+      after(() => runResult.cleanup());
+
+      it('should compose with jhipster:common', () => {
+        assert(runResult.mockedGenerators['jhipster:common'].calledOnce);
+      });
+      it('should compose with jhipster:languages', () => {
+        assert.equal(runResult.mockedGenerators['jhipster:languages'].callCount, 1);
+      });
+    });
+
+    describe('without cypress', () => {
+      let runResult;
+      const options = { testFrameworks: [] };
+      before(() => {
+        return helpers
+          .create(generatorFile)
+          .withOptions({
+            fromCli: true,
+            skipInstall: true,
+            defaultLocalConfig: { ...appDefaultConfig, ...options },
+          })
+          .withMockedGenerators(mockedComposedGenerators)
+          .run()
+          .then(result => {
+            runResult = result;
+          });
+      });
+
+      after(() => runResult.cleanup());
+
+      it('should compose with jhipster:common', () => {
+        assert(runResult.mockedGenerators['jhipster:common'].calledOnce);
+      });
+      it('should compose with jhipster:languages', () => {
+        assert(runResult.mockedGenerators['jhipster:languages'].calledOnce);
+      });
+      it('should not compose with jhipster:cypress', () => {
+        assert.equal(runResult.mockedGenerators['jhipster:cypress'].callCount, 0);
+      });
+    });
+
+    describe('with cypress', () => {
+      let runResult;
+      const options = { testFrameworks: [CYPRESS] };
+      before(() => {
+        return helpers
+          .create(generatorFile)
+          .withOptions({
+            fromCli: true,
+            skipInstall: true,
+            defaultLocalConfig: { ...appDefaultConfig, ...options },
+          })
+          .withMockedGenerators(mockedComposedGenerators)
+          .run()
+          .then(result => {
+            runResult = result;
+          });
+      });
+
+      after(() => runResult.cleanup());
+
+      it('should compose with jhipster:common', () => {
+        assert(runResult.mockedGenerators['jhipster:common'].calledOnce);
+      });
+      it('should compose with jhipster:languages', () => {
+        assert(runResult.mockedGenerators['jhipster:languages'].calledOnce);
+      });
+      it('should compose with jhipster:cypress', () => {
+        assert(runResult.mockedGenerators['jhipster:cypress'].calledOnce);
+      });
+    });
+  });
 });
